Extract CardBadges into card_badges.dart
