+++ lib/src/sdk/cards/cards.dart
@@ -1,6 +1,7 @@
 export 'attachment_client.dart';
 export 'attachment_id.dart';
 export 'card.dart';
+export 'card_badges.dart';
 export 'card_client.dart';
 export 'card_id.dart';
 export 'card_label.dart';
