+++ lib/src/sdk/cards/trello_card.dart
@@ -1,5 +1,6 @@
 import '../coordinates.dart' show Coordinates;
 import '../trello_object.dart' show TrelloObject;
+import 'card_badges.dart' show CardBadges;
 import 'card_id.dart' show CardId;
 import 'card_label.dart' show CardLabel;
 import 'card_models.dart' show CardBadges, CardFields;
